labs/03-rag/demo-03-embedding-with-chromadb/pyproject.toml
[project]
name = "demo-03-embedding-with-chromadb"
version = "0.1.0"
description = "Simple ChromaDB integration with OpenAI embeddings - Console application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.22",
    "python-dotenv>=1.0.0",
    "openai>=1.10.0"
]
